// src/app/model/inbox/AccountInfo.ts
export interface AccountInfo {
  Namedinsured_Fullname: string;
  Namedinsured_Mailingaddress_Lineone: string;
  Namedinsured_Mailingaddress_Cityname: string;
  Namedinsured_Mailingaddress_Stateorprovincecode: string;
  Namedinsured_Mailingaddress_Postalcode: string;
  Namedinsured_Naicscode: string;
  Naics_Description: string;
  Producer_Fullname: string;
  Insurer_Produceridentifier: string;
  Producer_Mailingaddress_Lineone: string;
  Producer_Mailingaddress_Postalcode: string;
}